+++ apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/claim/model/USD.java
@@ -1,11 +1,12 @@
 package gov.cms.bfd.server.ng.claim.model;
 
+import java.math.BigDecimal;
 import org.hl7.fhir.r4.model.Money;
 
 class USD extends Money {
   private USD() {}
 
-  static Money toFhir(double value) {
+  static Money toFhir(BigDecimal value) {
     return new Money().setCurrency("USD").setValue(value);
   }
 }
